requirements/local.txt
-r production.txt

Werkzeug[watchdog]==3.1.3 # https://github.com/pallets/werkzeug
ipdb==0.13.13  # https://github.com/gotcha/ipdb
psycopg[c]==3.2.3  # https://github.com/psycopg/psycopg

# Testing
# ------------------------------------------------------------------------------
mypy==1.13.0  # https://github.com/python/mypy
django-stubs[compatible-mypy]==5.1.1  # https://github.com/typeddjango/django-stubs
pytest==8.3.4  # https://github.com/pytest-dev/pytest
pytest-sugar==1.0.0  # https://github.com/Frozenball/pytest-sugar

# Documentation
# ------------------------------------------------------------------------------
sphinx==8.1.3  # https://github.com/sphinx-doc/sphinx
sphinx-autobuild==2024.10.3 # https://github.com/GaretJax/sphinx-autobuild

# Code quality
# ------------------------------------------------------------------------------
ruff==0.8.4  # https://github.com/astral-sh/ruff
coverage==7.6.9  # https://github.com/nedbat/coveragepy
djlint==1.36.3  # https://github.com/Riverside-Healthcare/djLint
pre-commit==4.0.1  # https://github.com/pre-commit/pre-commit

# Django
# ------------------------------------------------------------------------------
factory-boy==3.3.1  # https://github.com/FactoryBoy/factory_boy

django-debug-toolbar==4.4.6  # https://github.com/jazzband/django-debug-toolbar
django-extensions==3.2.3  # https://github.com/django-extensions/django-extensions
django-coverage-plugin==3.1.0  # https://github.com/nedbat/django_coverage_plugin
pytest-django==4.9.0  # https://github.com/pytest-dev/pytest-django

# My Addons
django-bootstrap-icons
